requirements.txt
python-telegram-bot=22.5
deluge-client=1.10.2